Fail on errors in test_container.sh
If one of the scripts in test_container.sh fails, or can't be run
we ended up reporting success. Change to bubble up the error, and
fail the entire test run.

We hit this problem with the hugectr tests being missing, but the
overall test_container.sh run looked successful.