Merlin offers three containers
In 22.05 and earlier, the NGC catalog provided one
container for training and a separate container for
inference.

For example, we had a `merlin-tensorflow-training` and
a `merlin-tensorflow-inference` container.

For 22.06, the training and inference software is combined
in a single container.  For example, the previous
capabilities are now provided in a `merlin-tensorflow`
container.