Add NGC overview descriptions for our containers
This adds the markdown text for our NGC overview section. The idea here is to
let us come together and figure out the best descriptions as a team, before
pushing out the description to NGC for the new containers.

This is a draft, merging the descriptions from the training/inference containers
together so that we can use with the unified containers