Address virtual review comments
This PR also includes updates that were triggered
by flake8_nb and also fixes to links based on the
output of `make -C docs linkcheck`.

Changes to input and output schema for Ensemble
in `test_scaling_criteo_merlin_models.py`
based on similar changes in the foll
PR that resembled the tests in _this_ repo.

https://github.com/NVIDIA-Merlin/systems/pull/197/

Importing merlin.models.tf is required.
Even though the import isn't used in the notebook
and is flagged by flake8-nb, importing the library
loads artifacts that are required so that Merlin
can run TensorFlow models.