docs: Add two projects to the support matrix (#720)
Add distributed_embeddings and merlin.dataloader.

Also add an override for the 22.10 HugeCTR version.
I don't want to keep updating that by hand.

Remove the Size value.  It became unreliable.
In 22.07 the value was ~14 G. In 22.09, it
became 220+ G and for 22.10, 500+ G.
Maybe something changed with our CI system.
Regardless, it's not as interesting as the
compressed size and that value stays.